Update tailwindcss config to include site
Before this commit the tailwindcss build would only consider the theme
files for building the style.css file.  This wasn't sufficient as we'll
want to include styling in our site content as well.

Update the tailwindcss config to also consider the site content files.

Also have it look for html files in addition to markdown files.