Update scbtheme head.html for SVG favicons
This isn't complete as I haven't tried to do anything special for Safari
pinned tab display (and it requires a "mask-icon" to be set).  The value
in doing that compared to my inability to test if I'm doing it right (no
Apple products or Safari to test with) makes costs outweigh benefits.