Fix some complaints from Edge developer tools
First, MS Edge browser didn't like the manifest.json file.  It
complained that it wasn't valid and that CORS was an issue.  The CORS
this likely would have resolved itself once the page was actually served
from a web server, but JSON validity wasn't clear.  I expect that it was
missing some field required by the standard, but rather than spend time
tracking that down... just kill it.  I'm not writing something that will
be a PWA and I'm not sure there's any other value to the manifest
outside of that context.

Next it was complaining about the charset.  This was almost certainly
because the preview server I'm using is injecting a <script> element
at the top of the <head> section, before any of my script's <meta>
elements.  Still, I tend to prefer utf-8 to UTF-8, so we'll keep the
change.