Remove the header.html partial from baseof.html
The base template is to fundamental to include the header.  There will
be times (possibly) when we don't want the header.  Other templates with
greater contextual awareness should be responsible for including the
header partial.

Tweak the style a bit for better presentation consistency (I believe
this was an issue for footers).