Update style.css based on current site.
The site is in a spot where it's looking fairly final, at least as far
as stylings are concerned.

Note that this is almost certainly the wrong approach though.  I have
both theme and non-theme styles here (to support working with tailwind
correctly), but the non-theme elements ending up in the theme
directories just isn't correct... it just works.

At some point we should figure out how to correctly build a theme css
and a site css and either have them both loaded or otherwise combined
into a single site based style.css.