Change chronology list style
Originally this list had no list style, but it was clear that some cue
that some entries were wrapped was necessary.  The circle style was
added on the thinking that it would add less weight to an already text-
heavy page... but it just didn't look right.  The less weight goal was
achieved, at the price of still not quite enough clarity.  Also, every
other list uses the "disc" style of bullet point so these lists looked
out of place in context to the site's style.

Change to disc to make the list clearer still and to maintain overall
site consistency.