Refactor single page layout to use baseof.html
Prior to this commit, the single.html pages in both the theme and as
defined in the site layouts required the full HTML header declaration
just as it appears in baseof.html in the default theme layouts.

This fix makes it so that both the theme and site level single.html
pages are calling the baseof.html so the header stuff need only be
defined once, rather than 3 times.

Found while adding Google Analytics... a head injected internal
template.